requirements.txt
pandas
numpy
matplotlib
seaborn
spotipy
mysql-connector-python
psychopg2-binary
python-dotenv